Fix playerMovement and recalculateStats
It appears that somewhere in development, the playerMovement call in Player.cs Update was tied to the Wit stat instead of the Speed stat, which makes finding defects with speed increases difficult.

Additionally, at some point, the assignment of charStatsBase appears to have been deleted, so we were having issues with null references when status effect and weapons effects came into play.

Both issues should be resolved with this commit.