Fix shadowing charStats, implement effect packets and destruction
There was an issue with the recalculateStats method where it was passed a copy of charStats (global), and then proceeded to work on charStats (internal/local) without returning it. In conjunction with using charStatsBase as the store for the original stats for each character, this made it impossible to remove stat boosts.
To fix this, charStatsBase was re-removed and the base character stats were made global to Player. recalculateStats is rewritten a bit to just use variables from the global scope rather than pass-ins and return an updated local copy of new stats that are built on top of the now global base stats.

The EffectPacket class is added so Players/Enemies can now pass indestructible copies of effects and duration counters around and not lose them between scene transitions.

Also, updateStatus is implemented on the player to check the expiration of effects and remove them if their durations expire.