Implement health effects on player, fix effect culling, update HealthManager
processHealthEffects is added to the player to process any status effects that require action on a per second basis (such as damage or healing over time). This is also added to the main Update method in Player

Fixed issue where persistent effects were culled immediately with a check for persistence.

Updated HealthManager to include a safe setter method for changing current HP, UpdateHealth. This prevents HP from going below 0 or above maxHP.