Add health effect processing to enemies, add weapon effect procc'ing
Update the EnemyHolder to process health-related status effects once per second. Also fix the EnemyHealth implementation on the EnemyHolder to reference the script attached to the editor Prefab rather than the empty scriptableobject (which is now just a variable).

Effects are also updated to be slightly more intuitive in that the percentage is 0-100 rather than 0-1 in the scriptableobjects

Also, enemy/weapon collisions now properly respond to the attackeffects attached to weapons. This means that DOTs and HOTs will actually proc and work as intended in conjunction with the above.